refactor(tool-call): colocate ToolCall with ToolCalls in adapters/types/tool.py
Moves ToolCall and to_tool_call back into dspy/adapters/types/tool.py
so the three tool-call types (Tool, ToolCall, ToolCalls) live in one
file. base_lm.py imports to_tool_call lazily inside its two hot paths
to avoid the base_lm → adapters/tool → base_type → base_lm import
cycle; base_type.py keeps BaseLM under TYPE_CHECKING for the same
reason.

Adds a TODO(MaximeRivest) marker on ToolCall and at the lazy-import
sites flagging that this interface should move to ToolPart, which
will let us collapse the lazy imports.

Public surface is unchanged: dspy.ToolCall, dspy.ToolCalls.ToolCall,
and dspy.adapters.{ToolCall,types.ToolCall} all resolve to the same
class object.
